partial backtrace_server refactor
